requirements.txt
mysql-connector-python
pyswip
ttkbootstrap
Pillow